refactor(milo): share one ListGrowthChart across Organizations/Projects/Users
The GrowthPoint/buildGrowthSeries/*GrowthChart trio was copy-pasted
near-verbatim across all three list pages, differing only in the item
type, its createdAt accessor, and the title string. Extracts a single
generic ListGrowthChart<T>({ items, getCreatedAt, title }) into
app/features/milo instead.
