+++ app/routes/customer/project/index.tsx
@@ -1,17 +1,15 @@
 import type { Route } from './+types/index';
 import { DateTime } from '@/components/date';
 import { DisplayId } from '@/components/display';
-import { DATE_RANGE_OPTIONS, ListPage, ListTable } from '@/features/milo';
+import { DATE_RANGE_OPTIONS, ListGrowthChart, ListPage, ListTable } from '@/features/milo';
 import { type GqlProject, useAllProjectsQuery } from '@/resources/request/client';
 import { orgRoutes } from '@/utils/config/routes.config';
 import { metaObject } from '@/utils/helpers';
 import { DataTable } from '@datum-cloud/datum-ui/data-table';
 import { t } from '@lingui/core/macro';
 import { createColumnHelper } from '@tanstack/react-table';
-import { format } from 'date-fns';
 import { useMemo } from 'react';
 import { Link } from 'react-router';
-import { Line, LineChart, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts';
 
 export const meta: Route.MetaFunction = () => {
   return metaObject(t`Projects`);
@@ -19,85 +17,7 @@ export const meta: Route.MetaFunction = () => {
 
 const columnHelper = createColumnHelper<GqlProject>();
 
-interface GrowthPoint {
-  month: string;
-  cumulative: number;
-}
-
-/** Buckets projects by creation month into a running total, for the growth chart. */
-function buildGrowthSeries(projects: GqlProject[]): GrowthPoint[] {
-  const created = projects
-    .map((p) => (p.createdAt ? new Date(p.createdAt) : null))
-    .filter((d): d is Date => d !== null && !Number.isNaN(d.getTime()))
-    .sort((a, b) => a.getTime() - b.getTime());
-
-  if (created.length === 0) return [];
-
-  const points: GrowthPoint[] = [];
-  let cumulative = 0;
-  let index = 0;
-  const cursor = new Date(created[0].getFullYear(), created[0].getMonth(), 1);
-  const end = new Date();
-
-  while (cursor <= end) {
-    const monthEnd = new Date(cursor.getFullYear(), cursor.getMonth() + 1, 1);
-    while (index < created.length && created[index] < monthEnd) {
-      cumulative++;
-      index++;
-    }
-    points.push({ month: format(cursor, 'MMM yyyy'), cumulative });
-    cursor.setMonth(cursor.getMonth() + 1);
-  }
-
-  return points;
-}
-
-/**
- * Cumulative-growth strip — rendered via `ListTable`'s `toolbar` slot, so it
- * sits in the table's own right-hand column (below the search bar, above the
- * column headers) rather than spanning over the filter sidebar.
- */
-function ProjectsGrowthChart({ projects }: { projects: GqlProject[] }) {
-  const growthData = useMemo(() => buildGrowthSeries(projects), [projects]);
-  const hasTrend = growthData.length >= 2;
-
-  return (
-    <div className="flex shrink-0 items-center gap-6 border-b px-4 py-3">
-      <div className="shrink-0">
-        <h2 className="text-muted-foreground text-sm font-medium">{t`Total projects`}</h2>
-        <span className="text-2xl font-semibold tabular-nums">{projects.length}</span>
-      </div>
-      <div className="min-w-0 flex-1">
-        {!hasTrend ? (
-          <div className="text-muted-foreground flex h-16 items-center text-sm">
-            {t`Not enough data yet to show a trend.`}
-          </div>
-        ) : (
-          <ResponsiveContainer width="100%" height={64}>
-            <LineChart data={growthData} margin={{ top: 4, right: 8, left: 0, bottom: 0 }}>
-              <XAxis
-                dataKey="month"
-                tickLine={false}
-                axisLine={false}
-                tick={{ fill: 'var(--muted-foreground)', fontSize: 11 }}
-                minTickGap={24}
-              />
-              <YAxis hide domain={['dataMin', 'dataMax']} />
-              <Tooltip labelFormatter={(month) => month} formatter={(value) => [value, t`Total`]} />
-              <Line
-                type="monotone"
-                dataKey="cumulative"
-                stroke="var(--primary)"
-                dot={false}
-                strokeWidth={2}
-              />
-            </LineChart>
-          </ResponsiveContainer>
-        )}
-      </div>
-    </div>
-  );
-}
+const getProjectCreatedAt = (project: GqlProject) => project.createdAt;
 
 export default function Page() {
   const tableQuery = useAllProjectsQuery();
@@ -150,7 +70,13 @@ export default function Page() {
         emptyMessage={t`No projects found.`}
         hasMore={tableQuery.data?.hasMore ?? false}
         hasMoreMessage={t`Limited to 10,000 projects. Refine your search to surface others.`}
-        toolbar={<ProjectsGrowthChart projects={projects} />}
+        toolbar={
+          <ListGrowthChart
+            items={projects}
+            getCreatedAt={getProjectCreatedAt}
+            title={t`Total projects`}
+          />
+        }
         filters={[
           { column: 'organizationName', label: t`Organization`, options: organizationOptions },
           {
