+++ app/routes/customer/user/index.tsx
@@ -4,7 +4,7 @@ import { DateTime } from '@/components/date';
 import { DialogForm } from '@/components/dialog';
 import { DisplayId, DisplayName, DisplayText } from '@/components/display';
 import { UserAvatar } from '@/components/user-avatar';
-import { DATE_RANGE_OPTIONS, ListPage, ListTable } from '@/features/milo';
+import { DATE_RANGE_OPTIONS, ListGrowthChart, ListPage, ListTable } from '@/features/milo';
 import { UserRejectDialog, useUserApproval } from '@/features/user';
 import {
   useAllUsersQuery,
@@ -22,10 +22,8 @@ import { t } from '@lingui/core/macro';
 import { Trans } from '@lingui/react/macro';
 import { ComMiloapisIamV1Alpha1User } from '@openapi/iam.miloapis.com/v1alpha1';
 import { createColumnHelper } from '@tanstack/react-table';
-import { format } from 'date-fns';
 import { useMemo, useState } from 'react';
 import { useNavigate } from 'react-router';
-import { Line, LineChart, ResponsiveContainer, Tooltip, XAxis, YAxis } from 'recharts';
 import { z } from 'zod';
 
 export const meta: Route.MetaFunction = () => {
@@ -34,85 +32,7 @@ export const meta: Route.MetaFunction = () => {
 
 const columnHelper = createColumnHelper<ComMiloapisIamV1Alpha1User>();
 
-interface GrowthPoint {
-  month: string;
-  cumulative: number;
-}
-
-/** Buckets users by creation month into a running total, for the growth chart. */
-function buildGrowthSeries(users: ComMiloapisIamV1Alpha1User[]): GrowthPoint[] {
-  const created = users
-    .map((u) => (u.metadata?.creationTimestamp ? new Date(u.metadata.creationTimestamp) : null))
-    .filter((d): d is Date => d !== null && !Number.isNaN(d.getTime()))
-    .sort((a, b) => a.getTime() - b.getTime());
-
-  if (created.length === 0) return [];
-
-  const points: GrowthPoint[] = [];
-  let cumulative = 0;
-  let index = 0;
-  const cursor = new Date(created[0].getFullYear(), created[0].getMonth(), 1);
-  const end = new Date();
-
-  while (cursor <= end) {
-    const monthEnd = new Date(cursor.getFullYear(), cursor.getMonth() + 1, 1);
-    while (index < created.length && created[index] < monthEnd) {
-      cumulative++;
-      index++;
-    }
-    points.push({ month: format(cursor, 'MMM yyyy'), cumulative });
-    cursor.setMonth(cursor.getMonth() + 1);
-  }
-
-  return points;
-}
-
-/**
- * Cumulative-growth strip — rendered via `ListTable`'s `toolbar` slot, so it
- * sits in the table's own right-hand column rather than spanning over the
- * filter sidebar. Mirrors the Projects/Organizations pages' growth charts.
- */
-function UsersGrowthChart({ users }: { users: ComMiloapisIamV1Alpha1User[] }) {
-  const growthData = useMemo(() => buildGrowthSeries(users), [users]);
-  const hasTrend = growthData.length >= 2;
-
-  return (
-    <div className="flex shrink-0 items-center gap-6 border-b px-4 py-3">
-      <div className="shrink-0">
-        <h2 className="text-muted-foreground text-sm font-medium">{t`Total users`}</h2>
-        <span className="text-2xl font-semibold tabular-nums">{users.length}</span>
-      </div>
-      <div className="min-w-0 flex-1">
-        {!hasTrend ? (
-          <div className="text-muted-foreground flex h-16 items-center text-sm">
-            {t`Not enough data yet to show a trend.`}
-          </div>
-        ) : (
-          <ResponsiveContainer width="100%" height={64}>
-            <LineChart data={growthData} margin={{ top: 4, right: 8, left: 0, bottom: 0 }}>
-              <XAxis
-                dataKey="month"
-                tickLine={false}
-                axisLine={false}
-                tick={{ fill: 'var(--muted-foreground)', fontSize: 11 }}
-                minTickGap={24}
-              />
-              <YAxis hide domain={['dataMin', 'dataMax']} />
-              <Tooltip labelFormatter={(month) => month} formatter={(value) => [value, t`Total`]} />
-              <Line
-                type="monotone"
-                dataKey="cumulative"
-                stroke="var(--primary)"
-                dot={false}
-                strokeWidth={2}
-              />
-            </LineChart>
-          </ResponsiveContainer>
-        )}
-      </div>
-    </div>
-  );
-}
+const getUserCreatedAt = (user: ComMiloapisIamV1Alpha1User) => user.metadata?.creationTimestamp;
 
 export default function Page() {
   const navigate = useNavigate();
@@ -343,7 +263,9 @@ export default function Page() {
           hasMoreMessage={t`Limited to 10,000 users. Refine your search to surface others.`}
           searchPlaceholder={t`Search users...`}
           emptyMessage={t`No users found.`}
-          toolbar={<UsersGrowthChart users={users} />}
+          toolbar={
+            <ListGrowthChart items={users} getCreatedAt={getUserCreatedAt} title={t`Total users`} />
+          }
           filters={[
             {
               column: 'status.registrationApproval',
